fix: remove warnings using shebang (#33)
supress node warnings as we are a CLI. No need to scare the users about
the experimental status of things.

patching emit didn't work. I thought it did, but it seems not. This
works, but i have no idea what windows will do with it.

_before_
```sh
❯ ./bin.js up ~/Pictures/3.jpg
  1 file (0.2MB)
⠋ Storing(node:39756) ExperimentalWarning: The Fetch API is an experimental feature. This feature could change at any time
(Use `node --trace-warnings ...` to show where the warning was created)
  bagbaieraytfm7hxlzzr4vmckazjiiwraozzrcj2bwqsipi5mtzgyxfc523lq
```

_after_
```sh
❯ ./bin.js up ~/Pictures/3.jpg
  1 file (0.2MB)
  bagbaieraytfm7hxlzzr4vmckazjiiwraozzrcj2bwqsipi5mtzgyxfc523lq
```

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>