feat: add support for sharded CAR uploads (#36)
Adds a `--car` flag to `w3 up` that allows the file to be a CAR file. If
flagged, we call `uploadCAR` on the client, which skips the unixfs
encoding but still shards and registers an upload.

Also adds a couple of parameters to control the shard size and request
concurrency: `--shard-size` and `--concurrent-requests`.

Finally, this also includes a fix to `filesize` which ensures it avoids
output like `0.0MB` and also calculates correctly the value for the
given units (we were dividing by 1024 not 1000 and should have been
outputting `MiB` if that were the case).