fix: pass cursor to list in `w3 ls`
fixes #48

I added a test for this, but the issue doesn't happen in our current mock impl. I can spend some more time to make the mock implementation support `--size` and `--cursor` but it's not clear to me that's the right move yet - feels a little like adding extra code just to make a  test fail.