feat: make email parameter to `space register` optional
If we're only authorized to act as a single account we shouldn't need to pass an email address to `space register` since we can infer which one to use from the delegations we currently have.

Preserve the ability to pass email for cases where we are authorized as multiple accounts.