feat: add ability to specify custom principal (#84)
This is needed by one of the teams integrating with w3up. 

w3up-client will already fail if you pass a different principal than the
store is configured to use, so we don't need to add any extra validation
logic - end-users can generate a key with `npx ucan-key ed --json` and
set the `W3_PRINCIPAL` environment variable for at least the first
invocation of `w3` in a new environment.

Setting `W3_PRINCIPAL` on subsequent invocations is optional but will
ensure the principal used by the client is always the same.