feat: add `name` to `w3 space info` output
It seemed odd that I have to use `w3 space ls` and scan the list to find the `*` to find the name of the currently selected space.

This fixes it so that `w3 space info` includes the name where possible.

The `space/info` capability doesn't provide the meta.name field of a space but perhaps it should.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>