wip: introduce `bridge generate-tokens` command
In service of the UCAN bridge implementation in https://github.com/web3-storage/w3infra/pull/325 this PR proposes a new `bridge generate-tokens` command that will generate values for the `Authorization` header and `proof` field of the HTTP request users make to the bridge.

This PR is currently intended as a stake-in-the-ground and is expected to change significantly.

TODO
- [ ] clean up and refactor bridge
- [ ] finalize command names
- [ ] decouple from coupon code (?)